test(devtools): fix transformInjectorResolutionPathsIntoTree tests
Previously transformInjectorResolutionPathsIntoTree returned an array, now it returns a tree node so we update the test cases to reflect that.
