+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/BUILD.bazel
@@ -1,6 +1,6 @@
 load("//devtools/tools:ng_module.bzl", "ng_module")
 load("@io_bazel_rules_sass//:defs.bzl", "sass_binary")
-load("//devtools/tools:typescript.bzl", "ts_test_library")
+load("//devtools/tools:typescript.bzl", "ts_library", "ts_test_library")
 load("//devtools/tools:defaults.bzl", "karma_web_test_suite")
 
 package(default_visibility = ["//:__subpackages__"])
@@ -21,6 +21,7 @@ ng_module(
     ],
     deps = [
         ":injector_providers",
+        ":injector_tree_fns",
         "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:injector_tree_visualizer",
         "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:resolution_path",
         "//devtools/projects/ng-devtools/src/lib/vendor/angular-split",
@@ -49,18 +50,29 @@ ng_module(
     ],
 )
 
-ts_test_library(
-    name = "injector_tree_test",
-    srcs = ["injector-tree.spec.ts"],
+karma_web_test_suite(
+    name = "test",
     deps = [
-        ":injector_tree",
+        ":injector_tree_fns_test_lib",
+    ],
+)
+
+ts_test_library(
+    name = "injector_tree_fns_test_lib",
+    srcs = [
+        "injector-tree-fns.spec.ts",
+    ],
+    deps = [
+        ":injector_tree_fns",
         "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:injector_tree_visualizer",
     ],
 )
 
-karma_web_test_suite(
-    name = "test",
+ts_library(
+    name = "injector_tree_fns",
+    srcs = ["injector-tree-fns.ts"],
     deps = [
-        ":injector_tree_test",
+        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:injector_tree_visualizer",
+        "//devtools/projects/protocol",
     ],
 )
