+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/injector-tree-fns.spec.ts
@@ -8,7 +8,7 @@
 
 import {InjectorTreeD3Node, InjectorTreeNode} from '../dependency-injection/injector-tree-visualizer';
 
-import {equalInjector, generateEdgeIdsFromNodeIds, getInjectorIdsToRootFromNode, grabInjectorPathsFromDirectiveForest, InjectorPath, splitInjectorPathsIntoElementAndEnvironmentPaths, transformInjectorResolutionPathsIntoTree} from './injector-tree.component';
+import {equalInjector, generateEdgeIdsFromNodeIds, getInjectorIdsToRootFromNode, grabInjectorPathsFromDirectiveForest, InjectorPath, splitInjectorPathsIntoElementAndEnvironmentPaths, transformInjectorResolutionPathsIntoTree} from './injector-tree-fns';
 
 describe('getInjectorIdsToRootFromNode', () => {
   it('should be able to get ids from a node', () => {
@@ -1715,37 +1715,1393 @@ describe('transformInjectorResolutionPathsIntoTree', () => {
       }
     ];
 
-    const expected: InjectorTreeNode[] = [{
-      'injector': {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+    const expected: InjectorTreeNode = {
+      'injector': {'name': '', 'type': 'hidden', 'id': 'N/A'},
       'children': [{
-        'injector': {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
-        'children': [
-          {
-            'injector': {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
-            'children': [{
-              'injector': {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+        'injector': {
+          'id': '1',
+          'type': 'element',
+          'name': '_AppComponent',
+          'node': {
+            'element': 'app-root',
+            'component': {'name': 'app-root', 'isElement': false, 'id': 0},
+            'directives': [],
+            'children': [
+              {
+                'element': 'router-outlet',
+                'component': null,
+                'directives': [{'name': '_RouterOutlet', 'id': 1}],
+                'children': [],
+                'resolutionPath': [
+                  {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                  {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                  {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                  {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                  {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                ]
+              },
+              {
+                'element': 'app-demo-component',
+                'component': {'name': 'app-demo-component', 'isElement': false, 'id': 2},
+                'directives': [],
+                'children': [
+                  {
+                    'element': 'router-outlet',
+                    'component': null,
+                    'directives': [{'name': '_RouterOutlet', 'id': 3}],
+                    'children': [],
+                    'resolutionPath': [
+                      {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                      {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                      {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                      {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                      {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                    ]
+                  },
+                  {
+                    'element': 'app-todo-demo',
+                    'component': {'name': 'app-todo-demo', 'isElement': false, 'id': 4},
+                    'directives': [],
+                    'children': [
+                      {
+                        'element': 'a',
+                        'component': null,
+                        'directives': [{'name': '_RouterLink', 'id': 5}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '8', 'type': 'element', 'name': '_RouterLink'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'a',
+                        'component': null,
+                        'directives': [{'name': '_RouterLink', 'id': 6}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '11', 'type': 'element', 'name': '_RouterLink'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'router-outlet',
+                        'component': null,
+                        'directives': [{'name': '_RouterOutlet', 'id': 7}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'app-todos',
+                        'component': {'name': 'app-todos', 'isElement': false, 'id': 8},
+                        'directives': [],
+                        'children': [
+                          {
+                            'element': 'a',
+                            'component': null,
+                            'directives': [{'name': '_RouterLink', 'id': 9}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '13', 'type': 'element', 'name': '_RouterLink'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          },
+                          {
+                            'element': 'a',
+                            'component': null,
+                            'directives': [{'name': '_RouterLink', 'id': 10}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '16', 'type': 'element', 'name': '_RouterLink'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          },
+                          {
+                            'element': 'a',
+                            'component': null,
+                            'directives': [{'name': '_RouterLink', 'id': 11}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '17', 'type': 'element', 'name': '_RouterLink'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          },
+                          {
+                            'element': 'app-todo',
+                            'component': {'name': 'app-todo', 'isElement': false, 'id': 12},
+                            'directives': [{'name': '_TooltipDirective', 'id': 13}],
+                            'children': [{
+                              'element': 'div',
+                              'component': null,
+                              'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                              'children': [],
+                              'resolutionPath': [
+                                {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                                {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                                {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                                {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                                {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                                {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                                {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                                {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                                {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                                {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                                {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                                {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                                {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                              ]
+                            }],
+                            'resolutionPath': [
+                              {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          },
+                          {
+                            'element': 'app-todo',
+                            'component': {'name': 'app-todo', 'isElement': false, 'id': 15},
+                            'directives': [{'name': '_TooltipDirective', 'id': 16}],
+                            'children': [{
+                              'element': 'div',
+                              'component': null,
+                              'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                              'children': [],
+                              'resolutionPath': [
+                                {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                                {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                                {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                                {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                                {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                                {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                                {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                                {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                                {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                                {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                                {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                                {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                                {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                                {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                              ]
+                            }],
+                            'resolutionPath': [
+                              {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          },
+                          {
+                            'element': '#comment',
+                            'component': null,
+                            'directives': [{'name': '_NgForOf', 'id': 18}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          }
+                        ],
+                        'resolutionPath': [
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      }
+                    ],
+                    'resolutionPath': [
+                      {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                      {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                      {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                      {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                      {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                      {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                    ]
+                  },
+                  {
+                    'element': 'app-heavy',
+                    'component': {'name': 'app-heavy', 'isElement': false, 'id': 20},
+                    'directives': [],
+                    'children': [],
+                    'resolutionPath': [
+                      {'id': '24', 'type': 'element', 'name': '_HeavyComponent'},
+                      {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                      {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                      {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                      {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                      {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                    ]
+                  }
+                ],
+                'resolutionPath': [
+                  {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                  {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                  {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                  {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                  {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                  {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                  {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                ]
+              }
+            ],
+            'resolutionPath': [
+              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+            ]
+          }
+        },
+        'children': [{
+          'injector': {
+            'id': '6',
+            'type': 'element',
+            'name': '_DemoAppComponent',
+            'node': {
+              'element': 'app-demo-component',
+              'component': {'name': 'app-demo-component', 'isElement': false, 'id': 2},
+              'directives': [],
               'children': [
                 {
-                  'injector': {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
-                  'children': [{
-                    'injector': {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
-                    'children': []
-                  }]
+                  'element': 'router-outlet',
+                  'component': null,
+                  'directives': [{'name': '_RouterOutlet', 'id': 3}],
+                  'children': [],
+                  'resolutionPath': [
+                    {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                    {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                    {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                    {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                    {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                  ]
                 },
                 {
-                  'injector': {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
-                  'children': [{
-                    'injector': {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
-                    'children': []
-                  }]
+                  'element': 'app-todo-demo',
+                  'component': {'name': 'app-todo-demo', 'isElement': false, 'id': 4},
+                  'directives': [],
+                  'children': [
+                    {
+                      'element': 'a',
+                      'component': null,
+                      'directives': [{'name': '_RouterLink', 'id': 5}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '8', 'type': 'element', 'name': '_RouterLink'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'a',
+                      'component': null,
+                      'directives': [{'name': '_RouterLink', 'id': 6}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '11', 'type': 'element', 'name': '_RouterLink'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'router-outlet',
+                      'component': null,
+                      'directives': [{'name': '_RouterOutlet', 'id': 7}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'app-todos',
+                      'component': {'name': 'app-todos', 'isElement': false, 'id': 8},
+                      'directives': [],
+                      'children': [
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 9}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '13', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 10}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '16', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 11}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '17', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'app-todo',
+                          'component': {'name': 'app-todo', 'isElement': false, 'id': 12},
+                          'directives': [{'name': '_TooltipDirective', 'id': 13}],
+                          'children': [{
+                            'element': 'div',
+                            'component': null,
+                            'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                              {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          }],
+                          'resolutionPath': [
+                            {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'app-todo',
+                          'component': {'name': 'app-todo', 'isElement': false, 'id': 15},
+                          'directives': [{'name': '_TooltipDirective', 'id': 16}],
+                          'children': [{
+                            'element': 'div',
+                            'component': null,
+                            'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                              {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          }],
+                          'resolutionPath': [
+                            {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': '#comment',
+                          'component': null,
+                          'directives': [{'name': '_NgForOf', 'id': 18}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }
+                      ],
+                      'resolutionPath': [
+                        {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                        {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    }
+                  ],
+                  'resolutionPath': [
+                    {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                    {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                    {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                    {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                    {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                    {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                  ]
+                },
+                {
+                  'element': 'app-heavy',
+                  'component': {'name': 'app-heavy', 'isElement': false, 'id': 20},
+                  'directives': [],
+                  'children': [],
+                  'resolutionPath': [
+                    {'id': '24', 'type': 'element', 'name': '_HeavyComponent'},
+                    {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                    {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                    {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                    {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                    {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                  ]
                 }
+              ],
+              'resolutionPath': [
+                {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                {'id': '4', 'type': 'null', 'name': 'Null Injector'}
               ]
-            }]
+            }
           },
-          {'injector': {'id': '24', 'type': 'element', 'name': '_HeavyComponent'}, 'children': []}
-        ]
+          'children': [
+            {
+              'injector': {
+                'id': '9',
+                'type': 'element',
+                'name': '_AppTodoComponent',
+                'node': {
+                  'element': 'app-todo-demo',
+                  'component': {'name': 'app-todo-demo', 'isElement': false, 'id': 4},
+                  'directives': [],
+                  'children': [
+                    {
+                      'element': 'a',
+                      'component': null,
+                      'directives': [{'name': '_RouterLink', 'id': 5}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '8', 'type': 'element', 'name': '_RouterLink'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'a',
+                      'component': null,
+                      'directives': [{'name': '_RouterLink', 'id': 6}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '11', 'type': 'element', 'name': '_RouterLink'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'router-outlet',
+                      'component': null,
+                      'directives': [{'name': '_RouterOutlet', 'id': 7}],
+                      'children': [],
+                      'resolutionPath': [
+                        {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    },
+                    {
+                      'element': 'app-todos',
+                      'component': {'name': 'app-todos', 'isElement': false, 'id': 8},
+                      'directives': [],
+                      'children': [
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 9}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '13', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 10}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '16', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'a',
+                          'component': null,
+                          'directives': [{'name': '_RouterLink', 'id': 11}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '17', 'type': 'element', 'name': '_RouterLink'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'app-todo',
+                          'component': {'name': 'app-todo', 'isElement': false, 'id': 12},
+                          'directives': [{'name': '_TooltipDirective', 'id': 13}],
+                          'children': [{
+                            'element': 'div',
+                            'component': null,
+                            'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                              {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          }],
+                          'resolutionPath': [
+                            {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': 'app-todo',
+                          'component': {'name': 'app-todo', 'isElement': false, 'id': 15},
+                          'directives': [{'name': '_TooltipDirective', 'id': 16}],
+                          'children': [{
+                            'element': 'div',
+                            'component': null,
+                            'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                            'children': [],
+                            'resolutionPath': [
+                              {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                              {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                              {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                              {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                              {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                              {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                              {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                              {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                              {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                              {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                              {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                              {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                              {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                            ]
+                          }],
+                          'resolutionPath': [
+                            {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        },
+                        {
+                          'element': '#comment',
+                          'component': null,
+                          'directives': [{'name': '_NgForOf', 'id': 18}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }
+                      ],
+                      'resolutionPath': [
+                        {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                        {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                        {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                        {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                        {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                        {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                        {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                        {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                        {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                        {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                      ]
+                    }
+                  ],
+                  'resolutionPath': [
+                    {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                    {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                    {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                    {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                    {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                    {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                  ]
+                }
+              },
+              'children': [{
+                'injector': {
+                  'id': '14',
+                  'type': 'element',
+                  'name': '_TodosComponent',
+                  'node': {
+                    'element': 'app-todos',
+                    'component': {'name': 'app-todos', 'isElement': false, 'id': 8},
+                    'directives': [],
+                    'children': [
+                      {
+                        'element': 'a',
+                        'component': null,
+                        'directives': [{'name': '_RouterLink', 'id': 9}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '13', 'type': 'element', 'name': '_RouterLink'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'a',
+                        'component': null,
+                        'directives': [{'name': '_RouterLink', 'id': 10}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '16', 'type': 'element', 'name': '_RouterLink'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'a',
+                        'component': null,
+                        'directives': [{'name': '_RouterLink', 'id': 11}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '17', 'type': 'element', 'name': '_RouterLink'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'app-todo',
+                        'component': {'name': 'app-todo', 'isElement': false, 'id': 12},
+                        'directives': [{'name': '_TooltipDirective', 'id': 13}],
+                        'children': [{
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }],
+                        'resolutionPath': [
+                          {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                          {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': 'app-todo',
+                        'component': {'name': 'app-todo', 'isElement': false, 'id': 15},
+                        'directives': [{'name': '_TooltipDirective', 'id': 16}],
+                        'children': [{
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }],
+                        'resolutionPath': [
+                          {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                          {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      },
+                      {
+                        'element': '#comment',
+                        'component': null,
+                        'directives': [{'name': '_NgForOf', 'id': 18}],
+                        'children': [],
+                        'resolutionPath': [
+                          {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      }
+                    ],
+                    'resolutionPath': [
+                      {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                      {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                      {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                      {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                      {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                      {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                      {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                      {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                      {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                      {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                    ]
+                  }
+                },
+                'children': [
+                  {
+                    'injector': {
+                      'id': '19',
+                      'type': 'element',
+                      'name': '_TodoComponent',
+                      'node': {
+                        'element': 'app-todo',
+                        'component': {'name': 'app-todo', 'isElement': false, 'id': 12},
+                        'directives': [{'name': '_TooltipDirective', 'id': 13}],
+                        'children': [{
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }],
+                        'resolutionPath': [
+                          {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                          {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      }
+                    },
+                    'children': [{
+                      'injector': {
+                        'id': '18',
+                        'type': 'element',
+                        'name': '_TooltipDirective',
+                        'node': {
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 14}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '18', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '19', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }
+                      },
+                      'children': []
+                    }]
+                  },
+                  {
+                    'injector': {
+                      'id': '22',
+                      'type': 'element',
+                      'name': '_TodoComponent',
+                      'node': {
+                        'element': 'app-todo',
+                        'component': {'name': 'app-todo', 'isElement': false, 'id': 15},
+                        'directives': [{'name': '_TooltipDirective', 'id': 16}],
+                        'children': [{
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }],
+                        'resolutionPath': [
+                          {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                          {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                          {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                          {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                          {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                          {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                          {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                          {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                          {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                          {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                          {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                          {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                        ]
+                      }
+                    },
+                    'children': [{
+                      'injector': {
+                        'id': '21',
+                        'type': 'element',
+                        'name': '_TooltipDirective',
+                        'node': {
+                          'element': 'div',
+                          'component': null,
+                          'directives': [{'name': '_TooltipDirective', 'id': 17}],
+                          'children': [],
+                          'resolutionPath': [
+                            {'id': '21', 'type': 'element', 'name': '_TooltipDirective'},
+                            {'id': '22', 'type': 'element', 'name': '_TodoComponent'},
+                            {'id': '20', 'type': 'element', 'name': '_NgForOf'},
+                            {'id': '14', 'type': 'element', 'name': '_TodosComponent'},
+                            {'id': '12', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '9', 'type': 'element', 'name': '_AppTodoComponent'},
+                            {'id': '5', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                            {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                            {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                            {'id': '15', 'type': 'environment', 'name': '_HomeModule'},
+                            {'id': '10', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                            {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                            {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                            {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                          ]
+                        }
+                      },
+                      'children': []
+                    }]
+                  }
+                ]
+              }]
+            },
+            {
+              'injector': {
+                'id': '24',
+                'type': 'element',
+                'name': '_HeavyComponent',
+                'node': {
+                  'element': 'app-heavy',
+                  'component': {'name': 'app-heavy', 'isElement': false, 'id': 20},
+                  'directives': [],
+                  'children': [],
+                  'resolutionPath': [
+                    {'id': '24', 'type': 'element', 'name': '_HeavyComponent'},
+                    {'id': '6', 'type': 'element', 'name': '_DemoAppComponent'},
+                    {'id': '0', 'type': 'element', 'name': '_RouterOutlet'},
+                    {'id': '1', 'type': 'element', 'name': '_AppComponent'},
+                    {'id': '7', 'type': 'environment', 'name': '_DemoAppModule'},
+                    {'id': '2', 'type': 'environment', 'name': '_AppModule'},
+                    {'id': '3', 'type': 'environment', 'name': 'Platform: core'},
+                    {'id': '4', 'type': 'null', 'name': 'Null Injector'}
+                  ]
+                }
+              },
+              'children': []
+            }
+          ]
+        }]
       }]
-    }];
+    };
 
     expect(transformInjectorResolutionPathsIntoTree(injectorPaths)).toEqual(expected);
   });
